+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * **\[arez\]** Avoid comparisons of state in `ReactArezComponent.shouldComponentUpdate()` as state is unused in arez
   based react components and `ReactArezComponent.scheduleArezKeyUpdate()` has been updated for forcefully schedule
   a render which skips `ReactArezComponent.shouldComponentUpdate()` lifecycle step.
+* **\[dom\]** Add missing "bottom" property to `react4j.dom.proptypes.html.CssProps`.
 
 ### [v0.95](https://github.com/react4j/react4j/tree/v0.95) (2018-09-06)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.94...v0.95)